Lots of refactoring
Routes no longer have stops or paths. They are simply metadata that holds basic things like the route name and agency. Everything that was in Routes are now in TripRoutes. A TripRoute is a specific path of a route. Typically, there are at least two TripRoutes, and inbound and outbound, however if the route has variations, then there is a TripRoute for variations. Trips are still the individual trips along a specific TripRoute.

This still isn't complete, but makes things fit the GTFS model mouch more.